requirements.txt
click
pytest
kubernetes
pyyaml
